Add Peridocity PML
Before, the NaNs could have been caused by unfettered growth between the
i = 1 and i = Nx slots, which I suspect will be mitigated by the
addition of the pml in those slots. I only weakly suspect, however, that
this will remove NaNs.

Note that Maxim's codes seem to have been amenable to this addition -
not much needed to change from the previous version, as there was
already an unused pml slot in Eyx.