Extend Planewave
Now that I have found the coding source of the problem, I can attempt to
understand it and alleviate it. Without the TF/SF boundary in the
x-direction, spurious Ey values are appearing at the edges of what
should be a planewave. I'm not sure how this affects the results.

In order to alleviate this issue, I will push the planewave to the
borders of the pml.